Allow auth by sessionid
Since early 2023 GaiaGPS now enforces human-ness in the web login
path by hidden re-captcha. This breaks user/pass authentication in
gaiagpsclient (intentionally). This change adds the ability to provide
the sessionid cookie via --sessionid at the command-line. This will
be persisted in the cookiejar in the same way, so it is only required
to be done initially and whenever the session expires.

Fixes #10